refactor: cast trait
